pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "deepguard"
version = "0.1.0"
description = "An Interactive Audio-Visual Agent for Deepfake Detection and Cognitive Intervention"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",

    # Audio processing
    "librosa>=0.10.0",
    "soundfile>=0.12.0",

    # Visual processing
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",

    # LLM integration
    "openai>=1.0.0",

    # Web interface
    "gradio>=4.0.0",

    # Utilities
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
mfa = [
    "montreal-forced-aligner>=3.0.0",
]

[project.scripts]
deepguard = "deepguard.app:main"

[tool.setuptools.packages.find]
include = ["deepguard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100
